Minor refactoring of steps
Steps no longer have a .Next property because there's no reason for a
step to be able to do something before and after the next step or to
prevent the execution of the next step. Originally this was necessary
because of the Zip step which had to switch to a temp folder before
calling the next step (and then zipping after it), but since Zip is no
longer a step, we don't need this any more. The reason zip is no longer
a step is because steps now only influence the contents of the files,
they no longer touch the files themselves (on disk) nor influence
formatting or encoding, this is now done by Model (formatting, encoding)
and Storage (file read/write) classes.