Add json extension to files in human-readable mode
Most json files in the pbit have no extension, but if we want good
support from VSCode we must add the extension. This way VSCode will add
syntax highlighting, breadcrumbs, schema etc for them. It does this
even without the extension for small files, but for large files it seems
that it doesn't autodetect that it's json so we get poor editing support.