ExportDataSourcesTransform => RegisterDataSourcesTransform
This transform now registers the data sources in the same file. The
reason is because we change all of the M queries that load the data, so
that they reference the registered data sources instead of repeating
the information from the data sources. This will allow us to switch
the data sources by merging with a .bim file with just data sources.