JsonElementZone: fixing problem with finding replaced elements using JPath
The ElementSelector property might not be valid any more after replacing
the original json element. This happens in cases where the selector uses
properties of the target objects to filter. This is now forbidden (only
implicitly, I'm afraid) for JsonElementZone derived types. Instead, the
selector must target all elements by their location in the object
hierarchy and any filtering on properties must be done in the
FilterSelectedElements method. A separate
FilterSelectedProcessedElements method is defined by JsonElementZone
which checks if the mappingZone matches the current zone (this property
is set during RegisterPayloadLocation).

That way, we can always find the target elements. When reading original
objects we use the derived classes filter. When reading processed objects
we use the mappingZone property.