Refactoring + unit testing (for Rename and Detect modelRef operations)
- The ReportInfoNavTransformBase no longer forces using a visitor, it
now has virtual Process methods. Derived classes are free to implement
them without visitors. This is done to enable the visitors to be hidden
inside extension methods which expose the operations to other C# code
making them easier to use and unit test.
- The base visitor no longer accepts a Path, it is the responsibility of
the user/caller to include path in logs (either by wrapping the logger
or simply by logging the path before the call).
- The QueryParser no longer reads the glossary directly but accepts
an IDbInfoGetter instead. Test can use this to provide a
naming-convention-based implementation rather than having to prepare a
glossary.
- Minor fix to ParserResultValidator
- Adding tests for detection and mapping of model references