Minification: "__low_line__" multiplication issue
The parser was replacing underscore, which is a valid identifier
character, with "__low_line__" which itself contains underscores which
would be replaced next time, and so on...

The solution is to simply leave the underscore character unchanged. All
the other replacement strings (e.g. "__greater_than_sign__") only use
letters and underscores, so they won't cause this recursive replacement
 issue.