Fix: calculated ignored tables no longer lose datatype
Example scenario: table "*MINT CLTFU" only exists in WLT. When we read
and save SABP model, it messes up the datatype of the "*MINT CLTFU"
table. This table is ignored by the SABP customization, but when saving
the SABP model, we need to manually add it into the model so we don't
lose it when we save to the repo. We were cloning the original table and
then including it in the model to be saved, but the problem was that the
table.Clone() method did not preserve column types for calculated
columns. The solution was to clone using a different mechanism
(JsonSerilaizer).