Validating object/filesystem names
This is to prevent the situation where an object name results in an
invalid (or even worse - silently sanitized) file/folder path which
confuses Packer when saving data to a folder which already contains
files.

A specific scenario is naming a report page with a trailing dot, e.g.
"Page1.". The folder is silently saved as "Page1" by windows which
confuses packer when saving over that folder as it looks like the
existing folder "Page1" has been untouched by the save (Packer sees
"Page1." as being touched and does not know that windows silently
changed that to "Page1"). Since Packer crops out any untouched files at
the end, this resunts in the folder toggling in and out of existence on
subsequent save opeartions.